+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Remove FormItemJointGroup component; Add `isJoint` prop to FormItemGroup
 - Remove Extra spacing after FormItemGroup
 - Add `fields` and `onFieldsChange` prop to FormWrapper to allow form creation using an array of config
-- RadioGroup now uses BaseInputComponent as well
+- RadioGroup and Select now uses BaseInputComponent
 <!-- Sidebar improvements -->
 <!-- - Fix Sidebar footer isStick -->